+++ frontend/src/components/MapView/LeftPanel/AnticipatoryActionPanel/AnticipatoryActionDroughtPanel/index.tsx
@@ -38,7 +38,7 @@ import {
 } from 'utils/server-utils';
 import { getFormattedDate } from 'utils/date-utils';
 import { DateFormat } from 'utils/name-utils';
-import { PanelSize } from 'config/types';
+import { AnticipatoryAction, PanelSize } from 'config/types';
 import { StyledCheckboxLabel, StyledRadioLabel } from './utils';
 import { StyledSelect } from '../utils';
 import DistrictView from './DistrictView/index';
@@ -46,6 +46,7 @@ import HomeTable from './HomeTable';
 import HowToReadModal from '../HowToReadModal';
 import Timeline from './Timeline';
 import Forecast from './Forecast';
+import { useAnticipatoryAction } from '../useAnticipatoryAction';
 
 const isZimbabwe = safeCountry === 'zimbabwe';
 
@@ -67,6 +68,10 @@ function AnticipatoryActionDroughtPanel() {
   const classes = useStyles();
   const dispatch = useDispatch();
   const { t } = useSafeTranslation();
+  const { AAData, AAConfig } = useAnticipatoryAction(
+    AnticipatoryAction.drought,
+  );
+  console.log({ AAData, AAConfig });
   const monitoredDistricts = useSelector(AAMonitoredDistrictsSelector);
   const AAAvailableDates = useSelector(AAAvailableDatesSelector);
   const selectedDistrict = useSelector(AASelectedDistrictSelector);
