+++ frontend/src/components/MapView/LeftPanel/AnticipatoryActionPanel/AnticipatoryActionStormPanel/index.tsx
@@ -17,7 +17,7 @@ import {
   AADataSelector,
   setAAFilters,
 } from 'context/anticipatoryAction/AAStormStateSlice';
-import { PanelSize } from 'config/types';
+import { AnticipatoryAction, PanelSize } from 'config/types';
 import { updateDateRange } from 'context/mapStateSlice';
 import { useDefaultDate } from 'utils/useDefaultDate';
 import { getFormattedDate } from 'utils/date-utils';
@@ -26,13 +26,15 @@ import { DateFormat } from 'utils/name-utils';
 import HowToReadModal from '../HowToReadModal';
 import ActivationTrigger from './ActivationTriggerView';
 import { StyledSelect } from '../utils';
+import { useAnticipatoryAction } from '../useAnticipatoryAction';
 
 function AnticipatoryActionStormPanel() {
   const classes = useStyles();
   const dispatch = useDispatch();
   const { t } = useSafeTranslation();
+  const { AAData } = useAnticipatoryAction(AnticipatoryAction.storm);
   const AAAvailableDates = useSelector(AAAvailableDatesSelector);
-  const AAData = useSelector(AADataSelector);
+  // const AAData = useSelector(AADataSelector);
   const [howToReadModalOpen, setHowToReadModalOpen] = React.useState(false);
 
   const selectedDate = useDefaultDate('anticipatory_action_storm');
