+++ frontend/src/components/MapView/LeftPanel/index.tsx
@@ -73,143 +73,9 @@ const LeftPanel = memo(() => {
   const dispatch = useDispatch();
   const tabValue = useSelector(leftPanelTabValueSelector);
 
-  const selectedLayers = useSelector(layersSelector);
-  const map = useSelector(mapSelector);
-  const { updateHistory, appendLayerToUrl, removeLayerFromUrl } =
-    useUrlHistory();
-
   const classes = useStyles({ tabValue });
 
   const isPanelHidden = tabValue === Panel.None;
-
-  const AALayerInUrl = React.useMemo(
-    () =>
-      selectedLayers.find(x => AALayerIds.includes(x.id as AnticipatoryAction)),
-    [selectedLayers],
-  );
-
-  // Load AA config based on AA layer id
-  const AAConfig = React.useMemo(() => {
-    if (AALayerInUrl && isAnticipatoryActionLayer(AALayerInUrl.id)) {
-      return getAAConfig(AALayerInUrl.id as AnticipatoryAction);
-    }
-    return null;
-  }, [AALayerInUrl]);
-
-  const AAData = useSelector((state: RootState) =>
-    AAConfig ? AAConfig.dataSelector(state) : null,
-  );
-  const AAAvailableDates = useSelector((state: RootState) =>
-    AAConfig ? AAConfig.availableDatesSelector(state) : null,
-  );
-  const loadAAData = AAConfig ? AAConfig.loadAction : null;
-  const serverAvailableDates = useSelector(availableDatesSelector);
-
-  // Sync serverAvailableDates with AAAvailableDates when the latter updates.
-  React.useEffect(() => {
-    if (AAAvailableDates) {
-      const combinedAvailableDates = isWindowedDates(AAAvailableDates)
-        ? getAAAvailableDatesCombined(AAAvailableDates)
-        : AAAvailableDates;
-      const updatedCapabilities = AALayerIds.reduce(
-        (acc, layerId) => ({
-          ...acc,
-          [layerId]: combinedAvailableDates,
-        }),
-        { ...serverAvailableDates },
-      );
-      dispatch(updateLayersCapabilities(updatedCapabilities));
-    }
-    // To avoid an infinite loop, we only want to run this effect when AAAvailableDates changes.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [AAAvailableDates, dispatch]);
-
-  // navigate to AA tab when app originally load with AA layer in url
-  React.useEffect(() => {
-    // TODO: update  Object.keys(AAData).length === 0 condition with something more solid
-    // Move to AA tab when directly linked there
-    if (
-      !isAnticipatoryActionLayer(tabValue) &&
-      AALayerInUrl !== undefined &&
-      AAData
-    ) {
-      if (
-        AALayerInUrl.id === AnticipatoryAction.drought &&
-        isWindowEmpty(AAData, 'Window 1')
-      ) {
-        dispatch(setTabValue(Panel.AnticipatoryActionDrought));
-      } else if (AALayerInUrl.id === AnticipatoryAction.storm) {
-        dispatch(setTabValue(Panel.AnticipatoryActionStorm));
-      }
-    }
-  }, [AAData, AALayerInUrl, dispatch, tabValue]);
-
-  // Remove from url when leaving from AA tab
-  React.useEffect(() => {
-    if (
-      !isAnticipatoryActionLayer(tabValue) &&
-      tabValue !== Panel.None &&
-      AALayerInUrl !== undefined &&
-      !isWindowEmpty(AAData, 'Window 1')
-    ) {
-      toggleRemoveLayer(
-        AALayerInUrl,
-        map,
-        getUrlKey(AALayerInUrl),
-        dispatch,
-        removeLayerFromUrl,
-      );
-    }
-  }, [AAData, AALayerInUrl, dispatch, map, removeLayerFromUrl, tabValue]);
-
-  // fetch csv data when loading AA page
-  React.useEffect(() => {
-    if (!isAnticipatoryActionLayer(tabValue) || !loadAAData) {
-      return;
-    }
-    dispatch(loadAAData());
-  }, [dispatch, tabValue, loadAAData]);
-
-  // Add or switch AA layers in url
-  React.useEffect(() => {
-    if (!isAnticipatoryActionLayer(tabValue)) {
-      return;
-    }
-    const selectedLayerId = AALayerIds.find(x => x === tabValue);
-    if (!selectedLayerId) {
-      return;
-    }
-
-    const layer = LayerDefinitions[
-      selectedLayerId
-    ] as AnticipatoryActionLayerProps;
-    if (!layer || AALayerInUrl?.id === layer.id) {
-      return;
-    }
-
-    if (AALayerInUrl) {
-      toggleRemoveLayer(
-        AALayerInUrl,
-        map,
-        getUrlKey(AALayerInUrl),
-        dispatch,
-        removeLayerFromUrl,
-      );
-    }
-
-    const updatedUrl = appendLayerToUrl(
-      getUrlKey(layer),
-      selectedLayers,
-      layer,
-    );
-    updateHistory(getUrlKey(layer), updatedUrl);
-
-    // Reset startDate
-    dispatch(updateDateRange({ startDate: undefined }));
-
-    // url does not instantly update. updateHistory and appendLayerToUrl functions re-trigger useEffect, before selected layers is set
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [selectedLayers, tabValue, dispatch, AALayerInUrl]);
 
   const renderedChartsPanel = React.useMemo(() => {
     if (!areChartLayersAvailable) {
